pyproject.toml
[tool.poetry]
name = "nijika"
version = "0.1.0"
description = ""
authors = ["yamause <yamause.devnezumi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
